Optimize jack check by removing `lsmod` & `grep`
In this case, lsmod(8) is redundant, because it's already parsing
'/proc/modules'. Here, I cut out the middle-man that is lsmod(8), and
also avoid unnecessarily loading up grep(1) and the subshell for the
pipe between them.

Unfortunately, matching `*jack*` not very specific, but I mimicked the
original grep(1) REGEX to avoid unexpected issues. If you're wanting a
module whose name is 'jack', then use a pattern of `jack`, but if
you're wanting a module whose name _contains_ the string 'jack', then
use the glob pattern introduced in this commit (`*jack*`).